ref_data_uploader/zipcode/requirements.txt
pandas>=2.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
